requirements.txt
python-docx
nvidia-ml-py
requests
Pillow